+++ printing/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Use PageTheme in example
 - Save shared pdf in cache on Android
 - Implement macOS embedding support
+- Implement Flutter Web support
 
 ## 2.1.9